### 2018-04-23 #694 @whedon
[PRE REVIEW]: Discrete Laplace mixture model with applications in forensic genetics
**Submitting author:** @mikldk (<a href="http://orcid.org/0000-0002-0234-0266">Mikkel Meyer Andersen</a>)
**Repository:** <a href="https://github.com/mikldk/disclapmix" target ="_blank">https://github.com/mikldk/disclapmix</a>
**Version:** v1.7
