+++ .travis.yml
@@ -26,27 +26,30 @@ after_script: R -q -e 'tic::after_script()'
 
 # Header
 language: r
-sudo: false
-dist: trusty
+sudo: true
 cache: packages
-latex: false
 
 matrix:
   include:
-  - r: 3.1
-  - r: 3.2
-  - r: oldrel
-  - r: release
-    env:
-    - BUILD_PKGDOWN: true
-  - r: devel
+    - os: linux
+      dist: trusty
+      r: oldrel
+    - os: linux
+      dist: trusty
+      r: release
+      env: 
+        - BUILD_PKGDOWN: true
+    - os: linux
+      dist: trusty
+      r: devel
+    - os: osx
+      osx_image: xcode8.3
+      r: release
 
-#env
 env:
   global:
   - _R_CHECK_FORCE_SUGGESTS_=false
   - MAKEFLAGS="-j 2"
   - TRAVIS_CXXFLAGS="-Wall -Wextra -pedantic -Werror"
 
-#services
 services:
