### 2018-05-14 #694 @arfon
@wrathematics & @nhejazi - would you be willing to review this submission for JOSS?
