### 2018-05-20 #694 @arfon
@pjotrp or @brainstorm - would you be willing to edit this submission for JOSS? @nhejazi has kindly offered to review this submission.
