### 2018-05-20 #694 @arfon
> I can also take this genetics contribution. @brainstorm if you are OK with that.

Perhaps one of you could review 😉
