### 2018-05-22 #748 @mikldk
@pjotrp thank you for editing this and @nhejazi thank you for reviewing this.

I have gone through the checklist and I hope that I have addressed everything on it to make the review as smooth as possible. Should I have overlooked something, please let me know and I will address it as fast as possible.

Again, thank you for helping out on this one. Let me know if there's anything that I can help with.
