### 2018-05-30 #748 @nhejazi
Thanks all for setting everything up for this review. I've just had a chance to start the review process (and have begun checking off items in the above list). Generally speaking, everything looks good. I plan to go through the remaining items and complete an initial review by the end of this week.
