### 2018-06-01 mikldk/disclapmix#1 @nhejazi
[JOSS Review] DOIs in references
As per a requirement of the [ongoing review of this R package for JOSS](https://github.com/openjournals/joss-reviews/issues/748), a few minor additions are necessary in the file `paper.bib`, and any place else in the R package in which bibliographic information is stored. Primarily, DOIs should be added to each article that has been assigned one. For example, the first entry in `paper.bib` should look like the following, where the last two lines are new.

```
@article{AndersenDisclap2013,
	 Author = {Mikkel Meyer Andersen and Poul Svante Eriksen and Niels Morling},
	 Journal = {Journal of Theoretical Biology},
	 Pages = {39--51},
	 Title = {{The discrete Laplace exponential family and estimation of Y-STR haplotype frequencies}},
	 Volume = {329},
         Year = {2013},
         doi = {10.1016/j.jtbi.2013.03.009},
         url = {https://doi.org/10.1016/j.jtbi.2013.03.009}
}
```
