### 2018-06-01 #748 @nhejazi
As this review progresses, I will be opening issues for any major comments and just a single issue for various minor suggestions (so as not to pollute the issue queue of the package). @mikldk, please feel free to let me know in each of the issues if you have any questions about this. Currently, the only thing outstanding in this review is the addition of DOIs to references in the file `paper.bib`
