### 2018-06-01 mikldk/disclapmix#2 @nhejazi
[JOSS Review] Suggestions
This issue provides various suggestions for this package that came to my attention as part of the [review of this R package for JOSS](https://github.com/openjournals/joss-reviews/issues/748). Addressing these items is not required for the JOSS review but they are listed here for the sake of completeness:

- [x] In each vignette, the name of the data set is given as "`dane`" whereas it is in fact "`danes`". This is a minor misspelling but still worth correcting.
- [ ] In the introductory vignette, the plot method of `disclapmixfit` objects produces a plot with `ggplot2`; however, a plot using base is constructed by hand shortly below. It would be more aesthetically consistent to use just one graphics/plotting ecosystem.
- [x] The `print` method for `disclapmixfit` objects is particularly uninformative --- it would be better if this method provided some condensed form (perhaps a table?) of the information presented by the `summary` method, or, alternatively, simply aliased the `summary` method.
- [x] In a few places in the vignettes, code like `1:ncol(df)` appears (where `df` is a `data.frame`) --- it is generally recommended to use `seq_along(df)` in such cases. Again, this is minor and doesn't occur anywhere in the codebase itself as best as I can tell.
