### 2018-06-01 5d3db76
Adding DOI's to paper.ibi
