+++ paper.bib
@@ -4,7 +4,9 @@
 	Pages = {39--51},
 	Title = {{The discrete Laplace exponential family and estimation of Y-STR haplotype frequencies}},
 	Volume = {329},
-	Year = {2013}}
+	Year = {2013},
+  doi = {10.1016/j.jtbi.2013.03.009},
+  url = {https://doi.org/10.1016/j.jtbi.2013.03.009}}
 	
 @book{DJBwoe2,
 	Author = {Steele, CD and Balding, DJ},
@@ -13,36 +15,40 @@
 	Edition = {2nd ed.},
 	Publisher = {Wiley},
 	Title = {Weight of evidence for forensic {DNA} profiles},
-	Year = 2015}
+	Year = 2015,	
+	doi = {10.1002/9780470867693}}
 	
 @article{Caliebe2010,
 	Author = {Amke Caliebe and Arne Jochens and Michael Krawczak and Uwe R{\"o}sler},
-	Issn = {0022-5193},
 	Journal = {Journal of Theoretical Biology},
 	Number = {2},
 	Pages = {336-342},
 	Title = {{A Markov Chain Description of the Stepwise Mutation Model: Local and Global Behaviour of the Allele Process}},
 	Volume = {266},
-	Year = {2010}}
+	Year = {2010},
+  doi = {10.1016/j.jtbi.2010.06.033},
+  url = {https://doi.org/10.1016/j.jtbi.2010.06.033}}
 	
 	
 @article{AndersenDisclapCluster2014,
-   author = {Mikkel Meyer Andersen and Poul Svante Eriksen and Niels Morling},
-   title = {{Cluster analysis of European Y-chromosomal STR haplotypes using the discrete Laplace method}},
-   journal = {Forensic Science International: Genetics},
-   pages = {182-194},
-   volume = {11},
-   year = {2014}
-}
+  author = {Mikkel Meyer Andersen and Poul Svante Eriksen and Niels Morling},
+  title = {{Cluster analysis of European Y-chromosomal STR haplotypes using the discrete Laplace method}},
+  journal = {Forensic Science International: Genetics},
+  pages = {182-194},
+  volume = {11},
+  year = {2014},
+  doi = {10.1016/j.fsigen.2014.03.016},
+  url = {https://doi.org/10.1016/j.fsigen.2014.03.016}}
 
 @article{AndersenDisclapMixture2015,
-   author = {Mikkel Meyer Andersen and Poul Svante Eriksen and Helle Smidt Mogensen and Niels Morling},
-   title = {{Identifying the most likely contributors to a Y-STR mixture using the discrete Laplace method}},
-   journal = {Forensic Science International: Genetics},
-   pages = {76-83},
-   volume = {15},
-   year = {2015}
-}
+  author = {Mikkel Meyer Andersen and Poul Svante Eriksen and Helle Smidt Mogensen and Niels Morling},
+  title = {{Identifying the most likely contributors to a Y-STR mixture using the discrete Laplace method}},
+  journal = {Forensic Science International: Genetics},
+  pages = {76-83},
+  volume = {15},
+  year = {2015},
+  doi = {10.1016/j.fsigen.2014.09.011},
+  url = {https://doi.org/10.1016/j.fsigen.2014.09.011}}
 
 @article{YHRD15,
 	Author = {Willuweit, S and Roewer, L},
@@ -50,9 +56,11 @@
 	Date-Modified = {2017-02-26 05:51:00 +0000},
 	Journal = {Forensic Science International: Genetics},
 	Pages = {43-48},
[…]
