### 2018-06-01 mikldk/disclapmix#1 @mikldk
Sorry about this. Fixed in 5d3db765054c5b9850be4809874405a378b714f4 (I couldn't find DOI for some of the older references). Will you confirm before I close this?
