+++ vignettes/refs.bib
@@ -4,8 +4,10 @@
 	Pages = {39--51},
 	Title = {{The discrete Laplace exponential family and estimation of Y-STR haplotype frequencies}},
 	Volume = {329},
-	Year = {2013}}
-	
+	Year = {2013},
+  doi = {10.1016/j.jtbi.2013.03.009},
+  url = {https://doi.org/10.1016/j.jtbi.2013.03.009}}
+  
 @article{Hallenberg2005YchromosomeSH,
   title={Y-chromosome STR haplotypes in Danes.},
   author={Charlotte Hallenberg and Karsten Nielsen and Bo Simonsen and Juan Sanchez and Niels Morling},
@@ -16,12 +18,14 @@
 }
 
 @article{AndersenDisclapMixture2015,
-	Author = {Mikkel Meyer Andersen and Poul Svante Eriksen and Helle Smidt Mogensen and Niels Morling},
-	Journal = {Forensic Science International: Genetics},
-	Pages = {76-83},
-	Title = {{Identifying the most likely contributors to a Y-STR mixture using the discrete Laplace method}},
-	Volume = {15},
-	Year = {2015}}
+  author = {Mikkel Meyer Andersen and Poul Svante Eriksen and Helle Smidt Mogensen and Niels Morling},
+  title = {{Identifying the most likely contributors to a Y-STR mixture using the discrete Laplace method}},
+  journal = {Forensic Science International: Genetics},
+  pages = {76-83},
+  volume = {15},
+  year = {2015},
+  doi = {10.1016/j.fsigen.2014.09.011},
+  url = {https://doi.org/10.1016/j.fsigen.2014.09.011}}
 
 @article{AndersenDisclapIntroduction2013,
 	Author = {Mikkel Meyer Andersen and Poul Svante Eriksen and Niels Morling},
