+++ vignettes/mixtures.Rmd
@@ -20,7 +20,7 @@ This vignette shows how to use the R package `disclapmix` that implements the me
 
 # Mixture analysis
 
-We again use the Danish reference database [@Hallenberg2005YchromosomeSH] with $n = 185$ observations (male Y-STR haplotypes) at $r=10$ loci is available in the `dane` dataset. 
+We again use the Danish reference database [@Hallenberg2005YchromosomeSH] with $n = 185$ observations (male Y-STR haplotypes) at $r=10$ loci is available in the `danes` dataset. 
 Let us load the package as well as the data:
 
 ```{r}
@@ -33,7 +33,7 @@ To fit the model, we need one observation per row.
 This is done for example like this:
 
 ```{r}
-db <- as.matrix(danes[rep(1:nrow(danes), danes$n), 1:(ncol(danes)-1)])
+db <- as.matrix(danes[rep(seq_len(nrow(danes)), danes$n), seq_len(ncol(danes)-1)])
 str(db)
 ```
 
