### 2018-06-01 mikldk/disclapmix#2 @mikldk
Thank you for the suggestions. I hope that all these are addressed in the above commit.
