### 2018-06-01 #748 @mikldk
Thanks, @nhejazi . I have addressed your issues and good suggestions in the issues you have opened. If you find more, please just open additional issues.
