### 2018-06-02 #748 @nhejazi
Thanks for making these changes. Everything looks good to me --- I can now recommend accepting this manuscript and code for publication. @mikldk I think the only thing remaining is for you to generate a new Zenodo archive DOI for the changes reflected in the most recent commits to your R package. Once that's done, I think either @arfon or I can finish accepting using the new DOI.
