### 2018-06-02 #748 @mikldk
Thanks @nhejazi. What is the recommended practise - to make a 1.7.1/1.8 release and update version in JOSS, right? Since 1.7 is already on CRAN it would be misleading to reuse this version for the updated package.
