### 2018-06-02 #748 @nhejazi
I think that since there have been a few minor changes made to the package as part of this review, it would make sense to bump the version number --- exactly how you do that is up to you, but I would personally recommend v1.7.1, and save v1.8.0 for the next time a CRAN release is done. Once the version has been bumped, you can tag that commit / create a GitHub release, assign it a new DOI through Zenodo and post that here for us to use in finishing up acceptance of the manuscript.
