### 2018-06-02 #748 @pjotrp
Thanks @nhejazi for your solid review. I agree, we should publish the latest. Update the DOI and we are good to go.
