### 2018-06-02 d3ff9c2
Bump to version 1.7.1
