### 2018-06-02 #748 @mikldk
Thanks, @nhejazi and @pjotrp . Version bump to 1.7.1 happened in mikldk/disclapmix@d3ff9c2e4e2bcd9f93921066543a3c1543c9d1cf and gave release https://github.com/mikldk/disclapmix/releases/tag/1.7.1 which resulted in DOI: 10.5281/zenodo.1257887 (<https://doi.org/10.5281/zenodo.1257887>)
