### 2018-06-02 #748 @pjotrp
@arfon looks like we are ready to publish. The release number at the top of the issue tracker should be 1.7.1 instead of 1.7.
