+++ DESCRIPTION
@@ -11,7 +11,7 @@ Authors@R: c(
 Author: Mikkel Meyer Andersen [aut, cre], Poul Svante Eriksen [aut]
 Maintainer: Mikkel Meyer Andersen <mikl@math.aau.dk>
 Description: Make inference in a mixture of discrete Laplace distributions using the EM algorithm. This can e.g. be used for modelling the distribution of Y chromosomal haplotypes as described in [1, 2] (refer to the URL section).
-License: GPL (>= 2) | file LICENSE
+License: GPL-2 | file LICENSE
 LinkingTo: Rcpp, RcppProgress
 Imports: Rcpp (>= 0.11), disclap (>= 1.4), cluster (>= 1.14.4), MASS, stats, graphics, methods, utils
 Suggests: knitr, ggplot2, gridExtra, ggdendro, scales, seriation, fwsim,
